requirements.txt
fastapi[all]
mlflow==2.1.1
azure-storage-blob==12.14.1
azure-identity==1.12.0
python-dotenv